services/api/requirements.txt
fastapi==0.112.2
uvicorn==0.30.6
strawberry-graphql==0.231.0
pyyaml==6.0.2
python-multipart==0.0.9
jsonschema==4.23.0
pydantic==2.6.4
httpx==0.27.2
tenacity==9.0.0
python-dotenv==1.0.1
